[util] Move StatTracker class into the utils
